Draggable custom lib is working
