+++ bsr.ps1
@@ -2,7 +2,29 @@
 # set up file path variables
 $chromeFolder = "./browser-search-regex/chrome/"
 $buildChromeFolder = "./build/chrome/"
+$customLibFolder = "./browser-search-regex/custom_lib/"
 $manifestFile = "manifest.json"
+
+# create the manifest.json file
+New-Item -Path "$($buildChromeFolder)$($manifestFile)" -ItemType File -Force | Out-Null
+
+# convert the contents of manifest.json to a powershell object
+$manifestJSON = Get-Content "$($chromeFolder)$($manifestFile)" -raw | ConvertFrom-Json
+
+function IncludeTSPathsInManifest {
+  param($parentDir)
+  Get-ChildItem "$($parentDir)*" -Include *.ts, *.css -Exclude *.d.ts, *background.ts | 
+    ForEach-Object { 
+      $fileName = (Split-Path -Leaf -Resolve $_)
+      if ((Get-ChildItem $_ | Select-Object Extension).Extension -eq ".ts") {
+        $manifestJSON.content_scripts[0].js.Add("./$($fileName)".replace('.ts', '.js'))
+      }
+      elseif ((Get-ChildItem $_ | Select-Object Extension).Extension -eq ".css") {
+        $manifestJSON.content_scripts[0].css.Add("./$($fileName)")
+      }
+    } | 
+    Out-Null 
+}
 
 # create necessary build folder if it does not exist
 if (-not (Test-Path -Path $buildChromeFolder)) {
@@ -19,11 +41,6 @@ Get-ChildItem "$($chromeFolder)" -Exclude *.ts, *.json |
   ForEach-Object {
     Copy-Item $_ "$($buildChromeFolder)$(Split-Path -Leaf -Resolve $_)" -Recurse
   }
-# create the manifest.json file
-New-Item -Path "$($buildChromeFolder)$($manifestFile)" -ItemType File -Force | Out-Null
-
-# convert the contents of manifest.json to a powershell object
-$manifestJSON = Get-Content "$($chromeFolder)$($manifestFile)" -raw | ConvertFrom-Json
 
 # insure that the js and the css attributes are of type ArrayList
 $originalJS = $manifestJSON.content_scripts[0].js
@@ -32,17 +49,8 @@ $originalCSS = $manifestJSON.content_scripts[0].css
 $manifestJSON.content_scripts[0].css = New-Object -TypeName 'System.Collections.ArrayList'
 
 # loop through each .ts and .css file and add their paths to the manifest
-Get-ChildItem "$($chromeFolder)*" -Include *.ts, *.css -Exclude *.d.ts, *background.ts -Recurse | 
-  ForEach-Object { 
-    $fileName = (Split-Path -Leaf -Resolve $_)
-    if ((Get-ChildItem $_ | Select-Object Extension).Extension -eq ".ts") {
-      $manifestJSON.content_scripts[0].js.Add("./$($fileName)".replace('.ts', '.js'))
-    }
-    elseif ((Get-ChildItem $_ | Select-Object Extension).Extension -eq ".css") {
-      $manifestJSON.content_scripts[0].css.Add("./$($fileName)")
-    }
-  } | 
-  Out-Null 
+IncludeTSPathsInManifest -parentDir $chromeFolder
+
 $manifestJSON.content_scripts[0].js = $originalJS + $manifestJSON.content_scripts[0].js
 $manifestJSON.content_scripts[0].css = $originalCSS + $manifestJSON.content_scripts[0].css
